Finally convert this crap to use stdio.  Apparently somebody did not know how to get the size of a file using stdio?  Or didn't know that using open(2) and friends is non-portable and dumb?  Who knows...
Also, more importantly, put the ACTUAL size of the file in the header,
not what we think it might possibly maybe be.  This is necessary for
PNCC features which do VAD (in fact doing VAD in feature extraction is
sort of generally useful...) since the actual size might be less than
the hypothetical maximum size.